Support for older dpkg
Prior to dpkg 1.16.2, dpkg-query did not support the binary:Package
virtual field. Where supported it provides the architecture
identifier as part of the package name for multi-arch packages, but
bindep's current syntax lacks support for multi-arch discriminators
anyway. Instead use the Package field in the dpkg-query format
string which will work as expected on all dpkg versions.

This is needed to support running on Ubuntu 12.04.

Change-Id: I5438083eb02b24f0fda61ad91bc6bf4ca0ddaa13