Use exit code 2 for missing packages/bad versions
Exit with 2 for missing packages and bad versions of installed
packages, so as not to conflate exit code 1 which was and still is
also used to indicate general execution errors.

This commit breaks backward compatibility, noted for posterity.

Extra effort is not expended to maintain the specific nonzero exit
code originally used for identifying missing packages/bad versions.

Change-Id: Iaf0899015a0e46b661cc6efd654e2118a320827d