Don't exit 2 when --brief is used
Because scripts are often designed to abort early when commands exit
nonzero and since bindep can be very useful in scripts (e.g., to
feed lists of missing packages directly into a distro package
manager command line invocation), it's helpful to avoid exiting with
a nonzero value when the command is doing what is expected within
that context. Otherwise users are stuck following all bindep calls
with ||true or toggling errexit.

Brief mode is already intended for script-friendly use cases, so
overload it with this differing exit code behavior... specifically,
exit 0 unless an actual execution error is encountered as long as
-b/--brief is specified.

This commit breaks backward compatibility, noted for posterity.

Change-Id: I63a32bf42385442f7a62298d6c077659e7a4a233