Add Gentoo support to bindep
Equery is used to query Gentoo's package database.
An exit code of 3 means the package is not installed or it is an
invalid package atom (doesn't exist for example).  We are able to rely
on the exit code for package install status and just output the
package version, if it's installed.  We do not output revision
information as it is not needed here.

We also add support for gentoo categories in the package name by
allowing a '/' in the package name.  This is needed for package names
that exist in multiple categories.  For instance, curl could map to
'dev-haskell/curl' or 'net-misc/curl'.  Without this change we would not
be able to specify.

Change-Id: Id87038a5599824befae627d67a03f59b70e19257