Fixed bindep -b mode for version mismatch errors
Before the patch, we were silently returning return code = 1 while not
writing out package name with the version required, making
install-distro-packages.sh fail.

Note that to be able to use the output generated by the tool as an input
for RPM based package managers (yum, dnf, zypper, ...), rpm package
should be at least >= 4.13 [1]. This is needed because earlier versions
of rpm did not support OR operator that is now used to generate some
versioned constraints in brief mode, specifically to implement !=
operator.

[1] http://rpm.org/wiki/PackagerDocs/BooleanDependencies

Change-Id: Id8f5fef7ee7623da909af021bc3fb04965aead80