Consistently check subprocess output
In some places we used universal_newlines and in others we didn't. In
some places we decoded from utf8 and in others we didn't. Switch to not
using universal_newlines at all because this implies getting back a str
instance on python 3 not a binary string which can't be decoded. Then
consistently decode from utf8 in all cases.

Note that we don't appear to do any line splitting on newlines here
anyways so this shouldn't affect anything, also these all run on Linux
where you should get unix newlines anyways.

Change-Id: I8dc5ee0b14906324169def2cdc35fbcd4f776423