Use locale's getpreferredencoding instead of utf8
Before we were decoding utf8 everywhere but its possible some other
encoding is in use. Use locale.getprefferedencoding to get what is
preferred as that should be what subprocesses return to us. This is more
in line with the old python3 behavior we got from check_output if we
weren't going to decode ourselves.

Change-Id: I69aea068a3c504cea6e6659bd61433bcac454dc0