Fix emerge testcases
During the refactoring of the depends module
I74ebd432d8e1ab8b3433f0f55d5dfe3535019911
it was found that two test cases for emerge
used the wrong type as a return type in the
CalledProcessError (str instead of bytes).

This patch fixes this and also adds a check that the
appropriate operations can be executed also for emerge.

Change-Id: I127e72cbcefaef12f28f5eb986cdcc769d033bf7
Signed-off-by: Andreas Florath <andreas@florath.net>