Log an error when lsb_release cannot be executed
Right now, if lsb_release is not on the system, bindep just dumps stack
and indicates "No such file or directory". This gives no indication of
what the problem is or what to do to resolve it. This change logs an
error message about lsb_release being missing in that case.

Change-Id: I501647770478fe154d66cdb2c6adff48973dd23c