Move to argparse
I presume optparse was a python2.6 compat thing; but since it's stated
in the NEWS file that it's not supported for several years, I think we
can switch.  Even if someone *was* running on centos6 era, the odds
they won't/can't have python-argparse installed anyway are minimal, so
it's not like this is a completely hard break.

Change-Id: Ib933db547a811d533bdc95b4836354e959c0a6f5