Fix user profile display
Without this patch, issuing 'bindep --profiles' to list the
profiles on projects like loci would fail. This is due
to this perfectly valid line in the loci code [1].

There are indeed too many values to unpack in [2], as
the rule[1] would in this case be a list, not the usual
2-tuple.

This is a problem, because in the case of loci, it prevents
a user from predicting the list of packages to be installed
on the system in a read only fashion.

This patch solves the problem by ensuring the lists are
flattened first.

Next to that, the output list was showing platform details,
which should be discarded. Using the _partition function
allows to avoid repetitions.

[1]: https://github.com/openstack/loci/blob/8982c3ae717578eb8e565fd75e78a800970dbb25/bindep.txt#L13
[2]: https://github.com/openstack-infra/bindep/blob/cec3a7576bc7599f97a705a42e3572133b4891a1/bindep/depends.py#L276

Change-Id: Ied4ac178e20daf35f105e2acae0a2b3e11219a56