Fix tox python3 overrides
We want to default to running all tox environments under python 3, so
set the basepython value in each environment other than default one.

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We do not want to set the override once in testenv, because that
breaks the more specific versions used in default environments like
py35 and py36.

Also enables all supported version in default envlist but allows
skipping tests if interpreters are missing. This easy development and
does not affect CI which enforces specific environments on each job.

Change-Id: I62465239bd6d2ba2979027a6fae287e9ea5aba80
Closes-Bug:  #1801657
Co-Authored-By: Sorin Sbarnea <ssbarnea@redhat.com>