Allow mentioning tools paths in bindep
yum/dnf package managers do allow mentioning full path to executable
instead of rpm name for installing it.

This allows users to install needed tool regardless what package name
may provide it (which may depend on what repos they have installed or
enabled).

This allows 'package' names to start with slash as this is a requirement
for triggering this feature in yum/rpm.

Change-Id: I7e47ee956e9496d9f47a98f9595790797c8d1789